Fix incorrect column identifier bytes problem when renaming a column
patch by Ke Han; reviewed by Stefan Miklosovic and Brandon Williams for CASSANDRA-18956
