+++ CHANGES.txt
@@ -1,4 +1,5 @@
 3.0.31
+ * Fix incorrect column identifier bytes problem when renaming a column (CASSANDRA-18965)
  * Upgrade OWASP to 10.0.0 (CASSANDRA-19738)
 Merged from 2.2:
  * Add termin-8-jdk as a valid jdk8 candidate in the debian package (CASSANDRA-19752)
